Use correct type for texture handles

+++ ref_gl/gl_local.h
@@ -93,7 +93,7 @@ typedef struct image_s
 	int		upload_width, upload_height;	// after power of two and picmip
 	int		registration_sequence;		// 0 = free
 	struct msurface_s	*texturechain;	// for sort-by-texture world drawing
-	unsigned long		texnum;						// gl texture binding
+	GLuint texnum;						// gl texture binding
 	//int		detailtexnum;
 	float	sl, tl, sh, th;				// 0,0 - 1,1 unless part of the scrap
 	//qboolean	scrap;
